backend/pyproject.toml
[project]
name = "pixla"
version = "0.1.0"
description = "Pixel art generator with self-hosted AI models"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "pillow>=11.0.0",
    "numpy>=2.0.0",
    "torch>=2.0.0",
    "diffusers>=0.32.0",
    "transformers>=4.46.0",
    "accelerate>=0.34.0",
    "safetensors>=0.4.0",
    "peft>=0.14.0",
    "aiofiles>=24.0.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
